# ClientApp/src/client-shared/model/createUserDto.ts
/**
 * LouSoft Innovations sample app
 * LouSoft sample app description
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface CreateUserDto { 
    firstName: string;
    lastName: string;
    password: string;
}